Refactor puzzle 24 to return a long not a string
Refactor puzzle 24 to return the answer as a 64-bit integer instead of
as a string.

+++ src/ProjectEuler.Tests/Puzzles/Puzzle037Tests.cs
@@ -35,10 +35,10 @@ namespace MartinCostello.ProjectEuler.Puzzles
         [InlineData(3797, false, 379)]
         [InlineData(379, false, 37)]
         [InlineData(37, false, 3)]
-        public static void Puzzle037_Truncate_Returns_Correct_Result(int value, bool removeLeft, int expected)
+        public static void Puzzle037_Truncate_Returns_Correct_Result(long value, bool removeLeft, long expected)
         {
             // Act
-            int actual = Puzzle037.Truncate(value, removeLeft);
+            long actual = Puzzle037.Truncate(value, removeLeft);
 
             // Assert
             Assert.Equal(expected, actual);
